update a little  code

+++ services/contactsService.js
@@ -2,37 +2,45 @@ const Contacts = require("../models/contactsModel");
 
 const listContacts = async () => await Contacts.find();
 
-const getContactById = async (id) => {
+const getContactById = async (id, user) => {
   try {
-    return Contacts.findOne({ _id: id });
+    return Contacts.findOne({ _id: id, owner: user.id });
   } catch (error) {
+    console.error(error);
     return false;
   }
 };
 
-const addContact = ({ name, email, phone, favorite }) =>
-  Contacts.create({ name, email, phone, favorite });
+const addContact = ({ name, email, phone, favorite }, user) =>
+  Contacts.create({ name, email, phone, favorite, owner: user.id });
 
-const removeContact = (id) => {
+const removeContact = (id, user) => {
   try {
-    return Contacts.findByIdAndRemove({ _id: id });
+    return Contacts.findByIdAndRemove({ _id: id, owner: user.id });
   } catch (error) {
+    console.error(error);
     return false;
   }
 };
 
-const updateContact = (id, body) => {
+const updateContact = (id, body, user) => {
   try {
-    return Contacts.findByIdAndUpdate({ _id: id }, body, { new: true });
+    return Contacts.findByIdAndUpdate({ _id: id, owner: user.id }, body, {
+      new: true,
+    });
   } catch (error) {
+    console.error(error);
     return false;
   }
 };
 
-const updateStatusContact = (id, body) => {
+const updateStatusContact = (id, body, user) => {
   try {
-    return Contacts.findByIdAndUpdate({ _id: id }, body, { new: true });
+    return Contacts.findByIdAndUpdate({ _id: id, owner: user.id }, body, {
+      new: true,
+    });
   } catch (error) {
+    console.error(error);
     return false;
   }
 };
